fix a stupid compiler warning...
warning: assuming signed overflow does not occur when assuming that (X - c) > X is always false [-Wstrict-overflow]
#    define Q_ASSERT_X(cond, where, what) ((!(cond)) ? qt_assert_x(where, what,__FILE__,__LINE__) : qt_noop())

i hope i didnt break it.

+++ src/gui/EdgeItemBase.cpp
@@ -122,7 +122,8 @@ void EdgeItemSimple::adjustEdgePositioning()
                         QPointF(0, mpLabel->boundingRect().height()));
 
     // draw the arrow!
-    QLineF lastSegment(mPoints.at(mPoints.size()-2), mPoints.at(mPoints.size()-1));
+    QLineF lastSegment(mPoints.at((size_t)mPoints.size() - 2),
+                       mPoints.at((size_t)mPoints.size() - 1));
     double angle = std::acos(lastSegment.dx() / lastSegment.length());
     if(lastSegment.dy() >= 0)
         angle = 2 * M_PI - angle;
